.requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
datetime==6.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.0
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.0
fastar==0.11.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
nodeenv==1.10.0
numpy==2.4.4
passlib==1.7.4
prisma==0.15.0
pydantic==2.13.2
pydantic-core==2.46.2
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pygments==2.20.0
pyjwt==2.12.1
python-dotenv==1.2.2
python-multipart==0.0.26
pytz==2026.1.post1
pyyaml==6.0.3
rich==15.0.0
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.58.0
shellingham==1.5.4
starlette==1.0.0
tomlkit==0.14.0
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
zope-interface==8.3
